Restructured home page by creating a new body component
